+++ frontend/Idle/src/pages/optionPage.jsx
@@ -1,5 +1,5 @@
 import { styled } from "styled-components";
-import { useContext, useLayoutEffect, useRef, useState } from "react";
+import { useContext, useLayoutEffect, useState } from "react";
 import {
   ALL,
   CONVENIENCE,
@@ -13,23 +13,14 @@ import { carContext } from "../store/context";
 import { useNavigate, useParams } from "react-router-dom";
 import { getWithQueryAPI } from "../utils/api";
 import { PATH } from "../constant/path";
-import CategoryTabs from "../components/common/tabs/CategoryTabs";
 import OptionMain from "../components/optionPage/optionMain/OptionMain";
-import { ReactComponent as ArrowLogo } from "../assets/images/arrowOption.svg";
-import WhiteButton from "../components/common/buttons/WhiteButton";
-import BlueButton from "../components/common/buttons/BlueButton";
 import WarningModal from "../components/common/modals/WarningModal";
 import DefaultOptionButton from "../components/optionPage/optionSub/DefaultOptionButton";
 import DefaultOption from "../components/optionPage/defaultOption/DefaultOption";
-import palette from "../styles/palette";
-import OptionBoxContainer from "../components/optionPage/optionSub/OptionBoxContainer";
 import ConfusingTooltip from "../components/common/toolTips/ConfusingTooltip";
-
-const BLUR_STATUS = {
-  LEFT_NONE: 1,
-  RIGHT_NONE: -1,
-  BOTH_VISIBLE: 0,
-};
+import OptionConfirmContainer from "../components/optionPage/optionSub/OptionConfirmContainer";
+import OptionBoxContainer from "../components/optionPage/optionSub/OptionBoxContainer";
+import OptionTabContainer from "../components/optionPage/optionMain/OptionTabContainer";
 
 function navigateTo(car, navigate) {
   if (car.detail.engines.name === undefined) navigate("/detail/engines");
@@ -52,37 +43,14 @@ function OptionPage() {
   const [selectedOption, setSelectedOption] = useState("");
   const [optionData, setOptionData] = useState(cachedOptionData);
   const [currentPage, setCurrentPage] = useState(0);
-  const [blurState, setBlurState] = useState(BLUR_STATUS.LEFT_NONE);
   const [filteredData, setFilteredData] = useState([]);
   const [selectedFunction, setSelectedFunction] = useState("");
   const [tooltipState, setTooltipState] = useState(true);
   const [isWarningOpen, setIsWarningOpen] = useState(false);
   const [isDefaultModalOpen, setIsDefaultModalOpen] = useState(false);
   const tabs = [ALL, SAFETY, STYLE, PROTECTION, CONVENIENCE];
-  const scrollBar = useRef();
   const navigate = useNavigate();
   let selectedOptionIds = [];
-
-  useLayoutEffect(() => {
-    if (!scrollBar.current) {
-      return;
-    }
-    const getScrollState = () => {
-      const element = scrollBar.current;
-      if (element.scrollLeft === 0) {
-        setBlurState(BLUR_STATUS.LEFT_NONE);
-      } else if (element.scrollWidth === element.clientWidth + element.scrollLeft) {
-        setBlurState(BLUR_STATUS.RIGHT_NONE);
-      } else {
-        setBlurState(BLUR_STATUS.BOTH_VISIBLE);
-      }
-    };
-
-    scrollBar.current?.addEventListener("scroll", getScrollState);
-    return () => {
-      scrollBar.current?.removeEventListener("scroll", getScrollState);
-    };
-  }, [scrollBar.current]);
 
   useLayoutEffect(() => {
     selectedOptionIds = [];
@@ -126,109 +94,28 @@ function OptionPage() {
     setSelectedFunction(filteredData[0]?.functions[0]);
   }, [filteredData]);
 
-  function handleTabChange(direction) {
-    const currentIndex = tabs.indexOf(currentTab);
-
-    if (direction === "next") {
-      if (currentIndex !== -1 && currentIndex + 1 < tabs.length) {
-        navigate(`/option/${tabs[currentIndex + 1]}`);
-      } else {
-        if (car.getAllOptionChecked()) navigate("/bill");
-        setIsWarningOpen(true);
-      }
-    } else if (direction === "prev") {
-      if (currentIndex > 0) {
-        navigate(`/option/${tabs[currentIndex - 1]}`);
-      } else {
-        navigate("/color/interior");
-      }
-    }
-  }
-
-  function TabClicked(idx) {
-    navigate(`/option/${tabs[idx]}`);
-  }
-
-  function ArrowButtonClicked(direction) {
-    const element = scrollBar.current;
-    direction === "LEFT" ? (element.scrollLeft -= 200) : (element.scrollLeft += 200);
-  }
-
   setClickedOptionPage(true);
   return (
     <>
       <StWrapper>
-        <StTabContainer>
-          {tabs.map((item, idx) => (
-            <CategoryTabs
-              key={idx}
-              text={TRANSLATE[item]}
-              isClicked={item === currentTab}
-              onClick={() => TabClicked(idx)}
-            />
-          ))}
-        </StTabContainer>
-        <StContentsContainer>
-          <OptionMain
-            optionData={filteredData}
-            selectedOption={selectedOption}
-            currentPage={currentPage}
-            setCurrentPage={setCurrentPage}
-            setSelectedFunction={setSelectedFunction}
-            selectedFunction={selectedFunction}
-          />
-        </StContentsContainer>
+        <OptionTabContainer tabs={tabs} currentTab={currentTab} />
+        <OptionMain
+          optionData={filteredData}
+          selectedOption={selectedOption}
+          currentPage={currentPage}
+          setCurrentPage={setCurrentPage}
+          setSelectedFunction={setSelectedFunction}
+          selectedFunction={selectedFunction}
+        />
         <StBottomContainer>
           <StWrapper>
-            <ArrowLeftContainer $blurState={blurState}>
-              <ArrowLogo
-                alt="ArrowLogoImg"
-                onClick={() => {
-                  ArrowButtonClicked("LEFT");
-                }}
-              />
-            </ArrowLeftContainer>
-            <StContainer ref={scrollBar}>
-              <OptionBoxContainer
-                filteredData={filteredData}
-                selectedOption={selectedOption}
-                setSelectedOption={setSelectedOption}
-                setTooltipState={() => setTooltipState(false)}
-              />
-            </StContainer>
-            <ArrowRightContainer $blurState={blurState}>
-              <ArrowLogo
-                alt="ArrowLogoImg"
-                onClick={() => {
-                  ArrowButtonClicked("RIGHT");
-                }}
-              />
-            </ArrowRightContainer>
+            <OptionBoxContainer filteredData={filteredData} selectedOption={selectedOption} setSelectedOption={setSelectedOption} setTooltipState={setTooltipState} />
             <DefaultOptionButton onClick={() => setIsDefaultModalOpen(true)} />
             <StTooltipContainer onClick={() => setTooltipState(false)}>
               <StTooltip isActive={tooltipState} />
             </StTooltipContainer>
           </StWrapper>
-          <StConfirmContainer>
-            <StConfirmHeader>
-              <Title>{TRANSLATE[currentTab]} 선택</Title>
-              <Description>원하는 {TRANSLATE[currentTab]}을 선택해주세요.</Description>
-            </StConfirmHeader>
-            <StButtonContainer>
-              <WhiteButton
-                text={"이전"}
-                onClick={() => {
-                  handleTabChange("prev");
-                }}
-              />
-              <BlueButton
-                text={"다음"}
-                onClick={() => {
-                  handleTabChange("next");
-                }}
-              />
-            </StButtonContainer>
-          </StConfirmContainer>
+          <OptionConfirmContainer tabs={tabs} currentTab={currentTab} setIsWarningOpen={setIsWarningOpen} />
         </StBottomContainer>
       </StWrapper>
       {isWarningOpen ? (
@@ -250,71 +137,6 @@ function OptionPage() {
 
 export default OptionPage;
 
-const StContainer = styled.div`
-  position: relative;
-  display: flex;
-  align-items: flex-start;
-  gap: 8px;
-  width: 850px;
-  overflow: scroll;
-  &::-webkit-scrollbar {
-    height: 26px;
-    width: 0px;
-  }
-  &::-webkit-scrollbar-thumb {
-    height: 3px;
-    border-top: 10px solid ${palette.Grey_1};
-    border-bottom: 14px solid ${palette.Grey_1};
-    border-radius: 3px;
-    background: ${palette.NavyBlue_5};
-  }
-  &::-webkit-scrollbar-track {
-    background-color: ${palette.Grey_1};
-  }
-`;
-const ArrowRightContainer = styled.div`
-  position: absolute;
-  top: 0;
-  display: flex;
-  width: 50px;
-  height: 166px;
-  background: linear-gradient(270deg, #f6f6f6 0%, rgba(246, 246, 246, 0) 100%);
-  flex-shrink: 0;
-  align-items: center;
-  right: 170px;
-  z-index: 1;
-  svg {
-    position: absolute;
-    right: 0;
-  }
-  svg:hover {
-    cursor: pointer;
-  }
-  display: ${({ $blurState }) => ($blurState === BLUR_STATUS.RIGHT_NONE ? "none" : "")};
-`;
-
-const ArrowLeftContainer = styled.div`
-  position: absolute;
-  top: 0;
-  display: flex;
-  width: 50px;
-  height: 166px;
-  background: linear-gradient(270deg, #f6f6f6 0%, rgba(246, 246, 246, 0) 100%);
-  flex-shrink: 0;
-  align-items: center;
-  z-index: 1;
-  svg {
-    position: absolute;
-    right: 0;
-  }
-  svg:hover {
-    cursor: pointer;
-  }
-  transform: scaleX(-1);
-  left: -10px;
-  display: ${({ $blurState }) => ($blurState === BLUR_STATUS.LEFT_NONE ? "none" : "")};
-`;
-
 const StWrapper = styled.div`
   display: flex;
   align-items: center;
@@ -331,60 +153,6 @@ const StBottomContainer = styled.div`
   width: 1024px;
 `;
 
-const StConfirmContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 39px;
-  width: 154px;
-`;
-
-const StConfirmHeader = styled.div`
-  display: inline-flex;
-  flex-direction: column;
-  align-items: flex-start;
-  gap: 4px;
-`;
-
-const StButtonContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-`;
-
-const Title = styled.h1`
-  color: #222;
-  font-family: "Hyundai Sans Text KR";
-  font-size: 16px;
-  font-style: normal;
-  font-weight: 700;
-  line-height: 24px;
-  letter-spacing: -0.48px;
-`;
-
-const Description = styled.p`
-  color: #222;
-  font-family: "Hyundai Sans Text KR";
-  font-size: 13px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: 165%;
-  letter-spacing: -0.39px;
-  white-space: nowrap;
-`;
-
-const StTabContainer = styled.div`
-  position: absolute;
-  top: 68px;
-  left: 128px;
-  display: inline-flex;
-  align-items: flex-start;
-  gap: 24px;
-`;
-const StContentsContainer = styled.div`
-  position: absolute;
-  top: 110px;
-  left: 128px;
-`;
 
 const StTooltip = styled(ConfusingTooltip)``;
